+++ app/src/main/java/org/nanolx/securitybypasser/apps/Miitomo.java
@@ -18,14 +18,14 @@ public class Miitomo implements IXposedHookLoadPackage
 		XposedBridge.log("Nanolx Security Bypasser: Entering Miitomo... Ready... Set... GO!");
 		XposedHelpers.findAndHookMethod("jp.co.nintendo.SecureController",
 			param.classLoader,
-			"ahR",
+			"ZN",
 			new XC_MethodReplacement()
 			{
 				@Override
 				protected Object replaceHookedMethod(MethodHookParam param) throws Throwable
 				{
 					XposedBridge.log("Nanolx Security Bypasser: Nintendo is trying to stop us, but I won't let it happen!");
-					XposedBridge.log("Hooked SecureController.ahS())");
+					XposedBridge.log("Hooked SecureController.ZN())");
 					return null;
 				}
 			});
@@ -38,7 +38,7 @@ public class Miitomo implements IXposedHookLoadPackage
 
 		XposedHelpers.findAndHookMethod("jp.co.nintendo.SecureController",
 			param.classLoader,
-			"ahS",
+			"ZO",
 			new XC_MethodReplacement()
 			{
 				@Override
@@ -52,7 +52,7 @@ public class Miitomo implements IXposedHookLoadPackage
 
 		XposedHelpers.findAndHookMethod("jp.co.nintendo.SecureController",
 			param.classLoader,
-			"L",
+			"P",
 			Activity.class,
 			new XC_MethodReplacement()
 			{
@@ -60,14 +60,14 @@ public class Miitomo implements IXposedHookLoadPackage
 				protected Object replaceHookedMethod(MethodHookParam param) throws Throwable
 				{
 					XposedBridge.log("Nanolx Security Bypasser: Nintendo, I'm afraid I can't let you do that...");
-					XposedBridge.log("Hooked SecureController.L())");
+					XposedBridge.log("Hooked SecureController.P())");
 					return null;
 				}
 			});
 
-		XposedHelpers.findAndHookMethod("jp.co.nintendo.v",
+		XposedHelpers.findAndHookMethod("jp.co.nintendo.w",
 			param.classLoader,
-			"dp",
+			"cS",
 			boolean.class,
 			new XC_MethodReplacement()
 			{
@@ -75,7 +75,7 @@ public class Miitomo implements IXposedHookLoadPackage
 				protected Object replaceHookedMethod(MethodHookParam p1) throws Throwable
 				{
 					XposedBridge.log("Nanolx Security Bypasser: All you logcat output are belong to us!");
-					XposedBridge.log("Trying to catch v.dp(), logcat output should be available now!)");
+					XposedBridge.log("Trying to catch w.cS(), logcat output should be available now!)");
 					return null;
 				}
 			});